Umbennenung Dateien. Verschiebung von Funktionen. Kommentare erweitert.
