+++ MEKB_H0_Anlage/MEKB_H0_Anlage.csproj
@@ -92,20 +92,23 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Hauptform\Form1_ButtonCtrl.cs">
+    <Compile Include="Hauptform\Hauptform_ButtonCtrl.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Hauptform\Form1_Signale.cs">
+    <Compile Include="Hauptform\Hauptform_Signale.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Hauptform\Form1_LokStrg.cs">
+    <Compile Include="Hauptform\Hauptform_LokStrg.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Hauptform\Form1.cs">
+    <Compile Include="Hauptform\Hauptform.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Hauptform\Form1.Designer.cs">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <Compile Include="Hauptform\Hauptform.Designer.cs">
+      <DependentUpon>Hauptform.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Hauptform\Hauptform_Fahrstrassen.cs">
+      <SubType>Form</SubType>
     </Compile>
     <Compile Include="MenuFenster\Belegtmelder_Ueberwachung.cs">
       <SubType>Form</SubType>
@@ -113,10 +116,10 @@
     <Compile Include="MenuFenster\Belegtmelder_Ueberwachung.Designer.cs">
       <DependentUpon>Belegtmelder_Ueberwachung.cs</DependentUpon>
     </Compile>
-    <Compile Include="MenuFenster\LokSuche.cs">
+    <Compile Include="ZusatzFenster\LokSuche.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="MenuFenster\LokSuche.Designer.cs">
+    <Compile Include="ZusatzFenster\LokSuche.Designer.cs">
       <DependentUpon>LokSuche.cs</DependentUpon>
     </Compile>
     <Compile Include="Zusatz\Belegtmelder.cs" />
@@ -143,10 +146,10 @@
     <Compile Include="Hauptform\Z21_CallBacks.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Hauptform\Form1_Gleisplan.cs">
+    <Compile Include="Hauptform\Hauptform_Gleisplan.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Hauptform\Form1_Invokes.cs">
+    <Compile Include="Hauptform\Hauptform_Invokes.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="MenuFenster\InfoBox.cs">
@@ -169,15 +172,15 @@
       <DependentUpon>Z21_Einstellung.cs</DependentUpon>
     </Compile>
     <Compile Include="Z21_Bibliothek\Z21_Konstanten.cs" />
-    <Compile Include="MenuFenster\ZugSteuerpult.cs">
+    <Compile Include="ZusatzFenster\ZugSteuerpult.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="MenuFenster\ZugSteuerpult.Designer.cs">
+    <Compile Include="ZusatzFenster\ZugSteuerpult.Designer.cs">
       <DependentUpon>ZugSteuerpult.cs</DependentUpon>
     </Compile>
     <Compile Include="Zusatz\Weichen.cs" />
-    <EmbeddedResource Include="Hauptform\Form1.resx">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <EmbeddedResource Include="Hauptform\Hauptform.resx">
+      <DependentUpon>Hauptform.cs</DependentUpon>
       <SubType>Designer</SubType>
     </EmbeddedResource>
     <EmbeddedResource Include="MenuFenster\Belegtmelder_Ueberwachung.resx">
@@ -189,7 +192,7 @@
     <EmbeddedResource Include="MenuFenster\LokEditor.resx">
       <DependentUpon>LokEditor.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="MenuFenster\LokSuche.resx">
+    <EmbeddedResource Include="ZusatzFenster\LokSuche.resx">
       <DependentUpon>LokSuche.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
@@ -203,7 +206,7 @@
     <EmbeddedResource Include="MenuFenster\Z21_Einstellung.resx">
       <DependentUpon>Z21_Einstellung.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="MenuFenster\ZugSteuerpult.resx">
+    <EmbeddedResource Include="ZusatzFenster\ZugSteuerpult.resx">
       <DependentUpon>ZugSteuerpult.cs</DependentUpon>
     </EmbeddedResource>
     <None Include="Properties\Settings.settings">
